Youtube.automation.agent/AuraLyrics/requirements.txt
billboard.py
yt-dlp
moviepy
Pillow
numpy
requests
beautifulsoup4
mutagen
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
syncedlyrics
